Monster refactor of how RPMs are built
The new structure is:

The patches and extra files we used to have in topdir/SOURCES are now in
{hadoop,hive,pig}/{patches,files}.

There's a script called hadoop/clouderify_hadoop.py which is responsible for
taking a pristine hadoop distribution and patching it, adding our files, etc.

HadoopDist:supportedHadoopPackage does:
  - unpack a pristine hadoop
  - move in the hadoop/ directory as cloudera/ and run clouderify_hadoop
  - build a hadoop package/tarball with our modifications and enabled contribs
  - its outputs are the hadoop jars and a distro tarball

The hadoop RPM spec now does essentially the same steps, but instead of having
all of the SOURCES and PATCHES included, it just adds our hadoop/ directory
as a single tarball SOURCE, then accomplishes the same thing as the distro
target above.

The hive and pig packages are essentially unmodified except for some paths.

Another major change is that the targets now only build SRPMS, since we all run
on different development platforms, etc. It should be possible to build the SRPMs
on any host that can install rpmbuild (including debian)