KITCHEN-307. Add a new script called staging-report get status of archives
Description: This script gives the user an overview of ressources used and available.
It will detail which Elastic IPs are used or available as well as the hostname is available.
It will also list all instances started along with additional information such as who does it belong to,
its security group and its status (official archive.cloudera.com for instance)

Reason: Improvment, bug fixes