Changes to make an rc.cloudera.com
This is a hack to allow the nightly script to stage to a new server called rc.cloudera.com. This is in effect a fork of the nightly script which lives now in a different branch. This is evil and should be replaced after the u0 release with something more flexible.