CLOUDERA-BUILD. Update pristine src tarball pointer.
This is the tarball from github for the rc2 tag (which was what was
released: https://github.com/apache/hadoop-common/tags). It's a tar of
the source, not the actual release tarball because the actual release
tarball I've discovered is broken, eg missing files that are in the
source but don't make it in the generated tarball.