KITCHEN-4943 - Handle new Maven version artifacts
Adds a new option to bumpMavenVersion.groovy (--incremental) to have
it iterate over the list of components, update the selected
component's Maven version, build/deploy the new Maven artifacts, and
then move on to the next one. Also adds metadata to
jenkins_metadata.json for which components to do this for and
optionally what properties in pom.xml need to be modified.

This is to be used for the pre-branching prep, not during actual
branching.